LayoutTests/fast/xmlhttprequest/xmlhttprequest-responsetype-before-open.html

<!DOCTYPE html>
<html>
<head>
    <script src="../js/resources/js-test-pre.js"></script>
    <script>
        description('This tests that the XMLHttpRequest responseType attribute is modifiable prior to being in the OPENED state.');
        var xhr;

        xhr = new XMLHttpRequest();
        evalAndLog("xhr.responseType = 'document';");
        evalAndLog("xhr.open('GET', 'http://mydomain');");
        shouldBeEqualToString('xhr.responseType', 'document');
    </script>
    <script src="../js/resources/js-test-post.js"></script>
</head>
<body>
    <div id="description"></div>
    <div id="console"></div>
</body>
</html>
